Added: BiscuitDipper script
Added main script to handle the gameplay. It handles the player dipping the biscuit, increasing the points and decreasing the integrity of the biscuit as you hold it. When the integrity goes to 0 the biscuit breaks. It sets the current biscuit on start which we can later change to what the player selects. All values refresh when the biscuit breaks or the player stops breaking. The player gains crumbs (points) when they let go before it breaks. At the moment, they don't save or anything, they just go back to 0 when you stop playing.