Added: Basic shop system for biscuits
Added a basic shop system for the biscuits. The player can now buy bisuits by clicking the buttons on th left. If they already have it unlocked, it selects it as the current biscuit. If they don't, it checks if they have enough crumbs to unlock it. There is not yet a shop for the drinks. Also, only the console shows the player if they have unlocked or selected a biscuit. There are no visual coordinators yet.