Added: Biscuit button visuals
Added a script to handle the button visuals for the biscuit buttons. They are grey when locked, white when unlocked and unselected, and green when selected. These can of course be changed. It's kind of confusing I think. There is a function in the BiscuitButtonVisuals script that handles the Updating of the visuals. The BiscuitUIMangager has a function that goes through each button calling this function based on booleans. Then that function in the UIManager is called in the BiscuitManager script. Right now, it just handles the biscuit buttons, even though the BiscuitManager handles updating drinks too.

I'm not sure if I should move the updating drinks to another DrinkManager and copy the scripts but for drinks. This wouldn't be the best for expandability, but for a gamejam it seems quick and simple. I also don't know how I would make the same script be used for both biscuits and drinks. It just seems like a lot of duplicate code though so I don't know.