Added: Drink Manager and Score Manager ; Modified: Biscuit scripts
Added a DrinkManager script and moved the drink code from the BiscuitManager to this one. I want to separate the drink stuff and biscuit stuff from the same scripts unless needed. The BiscuitDipper obviously needs both. I made a new script to handle the score instead of inside the BiscuitManager. I modified most of the biscuit scripts and Unity inspector to work with this change. Next, I am going to try implement the shop for the drinks.